Move Node to separate file
Will eventually be moved to a separate package and is useful outside
of the RunCommand
