+++ cmd/influxd/run.go
@@ -5,8 +5,6 @@ import (
 	"fmt"
 	"io/ioutil"
 	"log"
-	"net"
-	"net/http"
 	"net/url"
 	"os"
 	"path/filepath"
@@ -16,7 +14,6 @@ import (
 	"time"
 
 	"github.com/influxdb/influxdb"
-	"github.com/influxdb/influxdb/admin"
 	"github.com/influxdb/influxdb/collectd"
 	"github.com/influxdb/influxdb/graphite"
 	"github.com/influxdb/influxdb/messaging"
@@ -37,156 +34,6 @@ func NewRunCommand() *RunCommand {
 	return &RunCommand{
 		node: &Node{},
 	}
-}
-
-type Node struct {
-	Broker   *influxdb.Broker
-	DataNode *influxdb.Server
-	raftLog  *raft.Log
-
-	hostname string
-
-	adminServer     *admin.Server
-	clusterListener net.Listener      // The cluster TCP listener
-	apiListener     net.Listener      // The API TCP listener
-	GraphiteServers []graphite.Server // The Graphite Servers
-	OpenTSDBServer  *opentsdb.Server  // The OpenTSDB Server
-}
-
-func (s *Node) ClusterAddr() net.Addr {
-	return s.clusterListener.Addr()
-}
-
-func (s *Node) ClusterURL() *url.URL {
-	// Find out which port the cluster started on
-	_, p, e := net.SplitHostPort(s.ClusterAddr().String())
-	if e != nil {
-		panic(e)
-	}
-
-	h := net.JoinHostPort(s.hostname, p)
-	return &url.URL{
-		Scheme: "http",
-		Host:   h,
-	}
-}
-
-func (s *Node) Close() error {
-	if err := s.closeClusterListener(); err != nil {
-		return err
-	}
-
-	if err := s.closeAPIListener(); err != nil {
-		return err
-	}
-
-	if err := s.closeAdminServer(); err != nil {
-		return err
-	}
-
-	for _, g := range s.GraphiteServers {
-		if err := g.Close(); err != nil {
-			return err
-		}
-	}
-
-	if s.OpenTSDBServer != nil {
-		if err := s.OpenTSDBServer.Close(); err != nil {
-			return err
-		}
-	}
-
-	if s.DataNode != nil {
-		if err := s.DataNode.Close(); err != nil {
-			return err
-		}
-	}
-
-	if s.raftLog != nil {
-		if err := s.raftLog.Close(); err != nil {
-			return err
-		}
-	}
-
-	if s.Broker != nil {
-		if err := s.Broker.Close(); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (s *Node) openAdminServer(port int) error {
-	// Start the admin interface on the default port
-	addr := net.JoinHostPort("", strconv.Itoa(port))
-	s.adminServer = admin.NewServer(addr)
-	return s.adminServer.ListenAndServe()
-}
-
-func (s *Node) closeAdminServer() error {
-	if s.adminServer != nil {
-		return s.adminServer.Close()
-	}
-	return nil
-}
-
-func (s *Node) openListener(desc, addr string, h http.Handler) (net.Listener, error) {
-	var err error
-	listener, err := net.Listen("tcp", addr)
-	if err != nil {
-		return nil, err
-	}
-	go func() {
-		err := http.Serve(listener, h)
-
-		// The listener was closed so exit
-		// See https://github.com/golang/go/issues/4373
-		if strings.Contains(err.Error(), "closed") {
-			return
-		}
-		if err != nil {
-			log.Fatalf("%s server failed to serve on %s: %s", desc, addr, err)
-		}
-	}()
-	return listener, nil
-
-}
-
-func (s *Node) openAPIListener(addr string, h http.Handler) error {
-	var err error
-	s.apiListener, err = s.openListener("API", addr, h)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (s *Node) closeAPIListener() error {
-	var err error
-	if s.apiListener != nil {
-		err = s.apiListener.Close()
-		s.apiListener = nil
-	}
-	return err
-}
-
-func (s *Node) openClusterListener(addr string, h http.Handler) error {
-	var err error
-	s.clusterListener, err = s.openListener("Cluster", addr, h)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (s *Node) closeClusterListener() error {
-	var err error
-	if s.clusterListener != nil {
-		err = s.clusterListener.Close()
-		s.clusterListener = nil
-	}
-	return err
 }
 
 func (cmd *RunCommand) ParseConfig(path string) error {
